fix: Eliminate 230+ compilation warnings
- Added crate-level lints for unused variables and dead code in stub implementations
- Fixed unused imports in audit and invoice routes
- Fixed unused variables in workflow service
- Added non-snake-case allowance for QuickBooks API field naming conventions
- All builds now compile cleanly with zero warnings
- All unit tests still passing

Affected crates:
- billforge-db
- billforge-api
- billforge-quickbooks
- billforge-billing
- billforge-vendor-mgmt
- billforge-reporting
- billforge-invoice-processing
- billforge-worker

+++ backend/crates/api/src/state.rs
@@ -185,7 +185,7 @@ impl AppState {
             .bind(name)
             .bind(description)
             .bind(queue_type)
-            .bind(*is_default as i32)
+            .bind(*is_default)
             .execute(pool)
             .await
             .ok();
@@ -236,7 +236,7 @@ impl AppState {
             .bind(queue_uuid)
             .bind(name)
             .bind(description)
-            .bind(*priority as i32)
+            .bind(*priority)
             .bind(conditions)
             .bind(assign_to)
             .execute(pool)
@@ -440,7 +440,7 @@ impl AppState {
             .bind(line_id)
             .bind(invoice_uuid)
             .bind(description)
-            .bind(*quantity as i32)
+            .bind(*quantity)
             .bind(*unit_price as i64)
             .bind(total as i64)
             .bind(notes)
